gerador-rad/requirements.txt
absl-py==0.9.0
academic==0.5.1
ago==0.0.93
aiohttp==3.6.2
alabaster==0.7.8
appdirs==1.4.3
apturl==0.5.2
argcomplete==1.8.1
argh==0.26.2
asgiref==3.2.10
astroid==2.3.3
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atari-py==0.2.6
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.4
awscli==1.18.172
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.9.1
bibtexparser==1.2.0
bleach==3.1.5
blessings==1.6
blinker==1.4
blis==0.4.1
botocore==1.19.12
bpython==0.18
Brlapi==0.7.0
bs4==0.0.1
cachetools==4.1.0
caffeine==2.9.4
catalogue==1.0.0
cchardet==2.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
chromedriver-binary==87.0.4280.20.0
ci-info==0.2.0
click==7.1.2
cloudpickle==1.3.0
cmake==3.17.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
configparser==5.0.1
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
curtsies==0.3.1
cycler==0.10.0
cymem==2.0.3
dbus-python==1.2.16
deap==1.3.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.1
dm-tree==0.1.5
docopt==0.6.2
docutils==0.15.2
dotmap==1.3.23
elasticsearch==7.9.1
entrypoints==0.3
etelemetry==0.2.2
ewmh==0.1.6
fake-useragent==0.1.11
fastcluster==1.1.28
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gast==0.3.3
gevent==1.3.5
ghclone==1.2.0
google==2.0.3
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
GPUtil==1.4.0
greenlet==0.4.13
grpcio==1.28.1
gym==0.17.2
h5py==2.10.0
hashids==1.2.0
hjson==3.0.2
html5lib==1.0.1
httplib2==0.14.0
hurry.filesize==0.9
hyperlink==20.0.1
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.4
itemadapter==0.2.0
itemloaders==1.0.3
itsdangerous==1.1.0
jedi==0.15.2
jericho==3.0.2
jieba3k==0.35.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-http-over-ws==0.0.8
jupyter-server==0.1.1
jupyterlab==2.2.0
jupyterlab-pygments==0.1.1
jupyterlab-server==1.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.1
kiwisolver==1.2.0
langdetect==1.0.8
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
logilab-common==1.4.3
louis==3.12.0
lxml==4.6.1
lz4==3.1.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.2.1
mccabe==0.6.1
meld==3.20.2
mementos==1.3.1
mistune==0.8.4
mock==3.0.5
more-itertools==4.2.0
mpmath==1.2.1
msgpack==0.6.2
multidict==4.7.6
murmurhash==1.0.2
mysql-connector-python==8.0.15
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
networkx==2.4
news-please==1.5.13
newspaper3k==0.2.8
nibabel==3.2.0
nipype==1.5.1
nltk==3.5
notebook==6.0.3
numexpr==2.7.2
numpy==1.18.4
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.4
pandocfilters==1.4.2
parsel==1.6.0
parso==0.5.2
pathlib==1.0.1
pathtools==0.1.2
patsy==0.5.1
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
plotly==4.8.1
powerlaw==1.4.6
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.11.3
prov==2.0.0
psutil==5.7.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybars3==0.9.7
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
PyDispatcher==2.0.5
pydot==1.4.1
pydotplus==2.0.2
pyflakes==2.1.1
pyglet==1.5.0
Pygments==2.6.1
PyGObject==3.36.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyMeta3==0.5.1
PyMuPDF==1.18.2
PyMySQL==0.10.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.1.7.4
pyRFC3339==1.1
pyrsistent==0.16.0
pysvn==1.9.9
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-docx==0.8.10
python-dotenv==0.14.0
python-igraph==0.8.3
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
pyxnat==1.4
PyYAML==5.3.1
pyzmq==19.0.1
QtAwesome==0.4.4
qtconsole==4.7.3
QtPy==1.9.0
queuelib==1.5.0
ray==0.8.5
rdflib==5.0.0
readability-lxml==0.8.1
redis==3.4.1
regex==2020.10.28
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
roman==2.0.0
rope==0.10.5
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
Scrapy==2.4.0
seaborn==0.11.1
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
sgmllib3k==1.0.0
simplejson==3.16.0
sip==4.19.21
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
spacy==2.2.4
Sphinx==1.8.5
spyder==3.3.6
spyder-kernels==0.5.2
sqlparse==0.3.1
srsly==1.0.2
ssh-import-id==5.10
statsmodels==0.12.2
systemd-python==234
tables==3.6.1
tabulate==0.8.7
TatSu==5.5.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
testresources==2.0.1
texttable==1.6.3
textworld==1.3.1
Theano==1.0.4
thinc==7.4.0
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.0.2
toml==0.10.1
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
traits==6.1.1
Twisted==20.3.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
voila==0.1.21
w3lib==1.22.0
wadllib==1.3.3
warcio==1.7.4
wasabi==0.6.0
watchdog==0.9.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
yarl==1.4.2
zipp==1.0.0
zope.interface==5.2.0
